fix: refactor and abstract components
