+++ components/flowchart/FlowchartVisualizerSelect.tsx
@@ -3,7 +3,7 @@ import {
   Instance,
   useBoxContextSandboxObject,
 } from '@components/box-page/box-context/sandbox-object';
-import FlowchartComponentSelect from '@components/flowchart/FlowchartComponentSelect';
+import { FlowchartComponentSelect } from '@components/flowchart/FlowchartComponentSelect';
 import { useSandboxComponents } from '@components/playground/SandboxComponentsProvider';
 import { errorFlowchartIncompatibleComponent } from '@constants/flowchart';
 import getUsedSlotsForAlias from '@utils/box-config/getUsedSlotsForAlias';
@@ -18,7 +18,7 @@ import { useCallback, useMemo } from 'react';
 
 import { useFilteredObjectOptions } from './useFilteredObjectOptions';
 
-export default function FlowchartVisualizerSelect({
+export function FlowchartVisualizerSelect({
   alias,
   className,
   hideLabel,
